Riemann working

+++ particle.c
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include "particle.h"
 
-Particle new_Particle(int mass, float x, float y, float z, float vx, float vy, float vz){
+Particle new_Particle(int mass, double x, double y, double z, double vx, double vy, double vz){
   Particle particle;
   particle.mass = mass;
   particle.p[0] = x;
@@ -18,5 +18,5 @@ Particle new_Particle(int mass, float x, float y, float z, float vx, float vy, f
 }
 
 void print_Particle(Particle particle){
-  printf("%f\t%f\t%f\t%f\t%f\t%f\t%f\n",particle.mass, particle.p[0],particle.p[1],particle.p[2],particle.v[0],particle.v[1],particle.v[2]);
+  printf("%le\t%le\t%le\t%le\t%le\t%le\t%le\n",particle.mass, particle.p[0],particle.p[1],particle.p[2],particle.v[0],particle.v[1],particle.v[2]);
 }
